# Make callsite cache invalidation much faster