--- a/tokio-trace/benches/no_subscriber.rs
+++ b/tokio-trace/benches/no_subscriber.rs
@@ -23,8 +23,9 @@ fn bench_no_span_no_subscriber(b: &mut Bencher) {
 
 #[bench]
 fn bench_1_atomic_load(b: &mut Bencher) {
+    // This is just included as a baseline.
     let n = test::black_box(1);
     use std::sync::atomic::{AtomicUsize, Ordering};
     let foo = AtomicUsize::new(1);
-    b.iter(|| (0..n).fold(0, |new, old| foo.load(Ordering::Relaxed) ^ new));
+    b.iter(|| (0..n).fold(0, |new, _old| foo.load(Ordering::Relaxed) ^ new));
 }