When a call site caches a filter evaluation for a particular piece of
metadata, the cached filter result will need to be evaluated if the
subscriber context changes. These subscribers may have different filters
which may return different results for the same metadata, so we must not
use the cached result from one subscriber when the current subscriber
has changed.

Currently, this invalidation is implemented by having the dispatcher
track a set of the hashes of all the callsites it has filtered since the
last subscriber switch. This has serious performance implementations.

This branch changes the approach to have the call-site store an ID of
the last subscriber that filtered it, and assigning a unique ID to every
new dispatcher that's created. By comparing the ID of the current
dispatcher with that of the last one to filter the subscriber, we can
determine if the cached filter invocation is still valid.

I've also written a quick benchmark to assess the overhead of creating a
span when there isn't currently a subscriber, compared to not creating a
span at all.

Here's the benchmark results before this change:
```
     Running target/release/deps/no_subscriber-4d55a6a1581e9283

running 3 tests
test bench_1_atomic_load         ... bench:           0 ns/iter (+/- 0)
test bench_no_span_no_subscriber ... bench:           0 ns/iter (+/- 0)
test bench_span_no_subscriber    ... bench:         113 ns/iter (+/- 17)
```

And after:
```
     Running target/release/deps/no_subscriber-4d55a6a1581e9283

running 3 tests
test bench_1_atomic_load         ... bench:           0 ns/iter (+/- 0)
test bench_no_span_no_subscriber ... bench:           0 ns/iter (+/- 0)
test bench_span_no_subscriber    ... bench:          31 ns/iter (+/- 3)
```

This is part of the work towards  #61.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>